Separated MCO and evaluation in two different plugins
